requirements.txt
pyspark==3.5.0
pyarrow==14.0.0
pyiceberg==0.5.1
pandas==2.1.4
polars==0.20.0
s3fs==2023.12.0
boto3==1.34.0
SQLAlchemy==2.0.23
apache-airflow==2.8.0
pyyaml==6.0.1
great-expectations==0.18.8
pydantic==2.5.0
structlog==23.2.0
tenacity==8.2.3
cryptography==41.0.8
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1